# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library API Documentation",
        "description": "API Documentation for the Library Application",
        "contact": {
            "email": "alexander@crosp.net"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Production Server"
        }
    ],
    "paths": {
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "List all books",
                "description": "Retrieve a list of all books in the library.",
                "operationId": "2e11e7c3d690b3cdefda4fc599a97abf",
                "responses": {
                    "200": {
                        "description": "List of books retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create a new book",
                "description": "Add a new book to the library. Only users with the librarian role can perform this action.",
                "operationId": "02c38a3823b835bcaddf64d168b72be0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "publisher",
                                    "author",
                                    "genre",
                                    "publication_date",
                                    "pages",
                                    "price"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "The Great Gatsby"
                                    },
                                    "publisher": {
                                        "type": "string",
                                        "example": "Scribner"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "F. Scott Fitzgerald"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "example": "Fiction"
                                    },
                                    "publication_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1925-04-10"
                                    },
                                    "pages": {
                                        "type": "integer",
                                        "example": 180
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15.99
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden, insufficient role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden, insufficient role"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid input data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get a single book",
                "description": "Retrieve details of a specific book by its ID.",
                "operationId": "482ca006573aeb3b3f11223fbae3d416",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the book to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Delete a book",
                "description": "Remove a book from the library. Only users with the librarian role can perform this action.",
                "operationId": "7d3f012acd91d9d393d46f54cf663472",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the book to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Book deleted successfully"
                    },
                    "403": {
                        "description": "Forbidden, insufficient role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden, insufficient role"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Books"
                ],
                "summary": "Update a book",
                "description": "Update details of an existing book. Only users with the librarian role can perform this action.",
                "operationId": "ff893897a2c27e2f4c9d0fa787e0c194",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the book to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "The Great Gatsby (Updated)"
                                    },
                                    "publisher": {
                                        "type": "string",
                                        "example": "Scribner"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "F. Scott Fitzgerald"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "example": "Classic Fiction"
                                    },
                                    "publication_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1925-04-10"
                                    },
                                    "pages": {
                                        "type": "integer",
                                        "example": 200
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 18.99
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden, insufficient role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden, insufficient role"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid input data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Book": {
                "required": [
                    "id",
                    "title",
                    "publisher",
                    "author",
                    "genre",
                    "publication_date",
                    "pages",
                    "price"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "The Great Gatsby"
                    },
                    "publisher": {
                        "type": "string",
                        "example": "Scribner"
                    },
                    "author": {
                        "type": "string",
                        "example": "F. Scott Fitzgerald"
                    },
                    "genre": {
                        "type": "string",
                        "example": "Fiction"
                    },
                    "publication_date": {
                        "type": "string",
                        "format": "date",
                        "example": "1925-04-10"
                    },
                    "pages": {
                        "type": "integer",
                        "example": 180
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 15.99
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Enter JWT token in the format **Bearer &lt;token&gt;**",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}